# Cache file reads