Now every *match_json_schema* call reloads the whole schema from disk. It may slow when there are many files (~1 second for each test example in our case).

This PR caches document store on the class level.