@seanpdoyle 2019-06-28
Could you provide more details in the commit message?

I'm very interested to learn more about the implications of these changes in your test suite, is that something that you've measured?