@dizer 2019-07-01
Hi @seanpdoyle 

You can try benchmarks at https://github.com/dizer/json_matchers_cache

```
Original, single document, 1000 tries:
0.99 sec
Original, linked document, 1000 tries:
5.68 sec

Patched, single document, 1000 tries:
0.52 sec
Patched, linked document, 1000 tries:
0.92 sec
```

For our requests tests suite (124 JSON schema files, 5047 lines) it saved about a minute.