@Bugagazavr 2021-02-25
Hi @seanpdoyle, any chance to get it merged? This feature is very helpful, with this patched version our specs working faster (3-4 minutes vs 20 minutes).

If you don't have time to resolve conflicts let me know, I create a new PR.