@artofhuman 2022-03-16
Hi @seanpdoyle. We also have same issues with speed in our tests after upgrade from 0.9 to 0.11 version. Could you merge this patch?